+++ web/src/components/schoolkeuze/wordmark.tsx
@@ -2,6 +2,7 @@
 
 import * as React from "react";
 import Link from "next/link";
+import Image from "next/image";
 import { cn } from "@/lib/utils";
 
 interface WordmarkProps {
@@ -11,10 +12,13 @@ interface WordmarkProps {
 function Wordmark({ className }: WordmarkProps) {
   return (
     <Link href="/" className={cn("inline-flex self-start", className)}>
-      <img
+      <Image
         src="/branding/mijnschoolkeuze_kit_v4/wordmark.png"
         alt="Mijn Schoolkeuze"
+        width={180}
+        height={36}
         className="h-9 w-auto object-contain"
+        priority
       />
     </Link>
   );